refactor: move MultiSigPubkeyHash to template folder structure
Matches our fork's organization pattern for templates.
