+++ package.json
@@ -11,20 +11,13 @@
     "dist/cjs/mod.js.map",
     "dist/cjs/package.json",
     "dist/cjs/src/template",
-    "dist/cjs/src/MultiSigPubkeyHash.js",
-    "dist/cjs/src/MultiSigPubkeyHash.js.map",
     "dist/esm/mod.js",
     "dist/esm/mod.js.map",
     "dist/esm/src/template",
-    "dist/esm/src/MultiSigPubkeyHash.js",
-    "dist/esm/src/MultiSigPubkeyHash.js.map",
     "dist/types/mod.d.ts",
     "dist/types/mod.d.ts.map",
     "dist/types/src/template",
-    "dist/types/src/MultiSigPubkeyHash.d.ts",
-    "dist/types/src/MultiSigPubkeyHash.d.ts.map",
     "src/template",
-    "src/MultiSigPubkeyHash.ts",
     "mod.ts",
     "LICENSE.txt"
   ],
